% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{barplt.UniChar}
\alias{barplt.UniChar}
\title{barplt.UniChar}
\usage{
barplt.UniChar(obj, ind, categ = FALSE)
}
\arguments{
\item{obj}{of Unichar class}

\item{ind}{an indicator to represent graphically. Must be a single column Data frame}

\item{categ}{default FALSE. Set true if your indicator was calculated on categorical data}
}
\value{
\item{bargraph}{ bar plot to show}
}
\description{
barplt.UniChar
}
