% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{circleFun}
\alias{circleFun}
\title{\code{circleFun} returns the x and y coordinates of a circle.}
\usage{
circleFun(center = c(0, 0), diameter = 2, npoints = 100)
}
\arguments{
\item{center}{Vector with the coordinates of the center. Default value is (0,0)}

\item{diameter}{Diameter of the circle. Default value is 2.}

\item{npoints}{Number of points used to create the circle. Default value is 100.}
}
\value{
data.frame with the coordinates of the points of the circle.
}
\description{
\code{circleFun} returns the x and y coordinates of a circle.
}
\examples{
circle <- circleFun()
plot(circle)

}
