% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{correlation_circle}
\alias{correlation_circle}
\title{Creation of the list of correlations and the list of squared correlations.}
\usage{
correlation_circle(X, data, clusters_name)
}
\arguments{
\item{X}{data.frame with the predictive variable that were used for the clustering}

\item{data}{List of data.frame for each clusters.}

\item{clusters_name}{Vector of the names of the clusters}
}
\value{
List with $list_corr (list of correlations) and $list_corr_carr (list of squared correlations).
}
\description{
This function is used in \code{multivariate_charac()}.
}
\examples{
require(readxl)
df<- read_excel('../Autos.xlsx') #specify the path
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
clus<- kmeans(X,3)
y <-clus$cluster


data<-list()
clusters_name<-sort(unique(y))
for (e in clusters_name){
  data=list.append(data,X[y==e,])
}

correlation_circle(X,data,clusters_name)

}
