% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{d}
\alias{d}
\title{Squared euclidean ditance}
\usage{
d(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
squared distance
}
\description{
Squared euclidean ditance
}
\examples{
x=c(0,1)
y=c(1,0)
dist=d(x,y)
}
