% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{effectsize.default}
\alias{effectsize.default}
\title{effectsize.default}
\usage{
\method{effectsize}{default}(obj, y)
}
\arguments{
\item{obj}{of UniChar class}

\item{y}{a vector of groups issued from a clustering process}
}
\value{
effect_sizes
}
\description{
Caracterisation function of groups with effect size indicator.
Applied on numerical features of obj.
}
\examples{
#Import data
df<- read_excel('Autos.xlsx')
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]

#Clustering
clus<- kmeans(X,3)
y <-clus$cluster

#Attribut de la classe UniChar
obj <- Perform_UniChar(df,y)

#Effect size
es <- effectsize(obj, y)
print(es)
}
