% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{indexes}
\alias{indexes}
\title{\code{indexes()}
Computing internal (and external) measures of a clustering}
\usage{
indexes(X, y, y_true = NULL)
}
\arguments{
\item{X}{data.frame with the predictive variable that were used for the clustering}

\item{y}{data.frame (or vector) with the same number of rows (or length) as \code{X}. This represents the predicted classes.}

\item{y_true}{data.frame (or vector) with the same number of rows (or length) as \code{X}. This represents the predicted classes.}
}
\value{
object of measures class - with internal and external clustering measures
}
\description{
\code{indexes()}
Computing internal (and external) measures of a clustering
}
\details{
measures objects have the following attributes :
\itemize{
\item{"internal"}{internal_ind object with the following measures as attributes:
\itemize{
\item{"Silhouette"}{Silhouette index.}
\item{"Dunn"}{Dunn index.}
\item{"eta_sq"}{Ratio of the intracorrelation on the total correlation.}
}
}

\item{"external"}{external_ind object with the following measures as attributes:
\itemize{rand_ind=rand_ind,accuracy=accuracy,kappa=kappa,F1=F1,recall=recall
\item{"rand_ind"}{Rand Index.}
\item{"Accuracy"}{Average accuracy of the clustering.}
\item{"kappa"}{Averaged value of kappa.}
\item{"F1"}{Averaged value of F1.}
\item{"recall"}{Average recall of the clustering.}
}
}
}
}
\examples{
require(readxl)
require(stats)
df<- read_excel('Autos.xlsx')
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
clus<- kmeans(X,3)
y <-clus$cluster

measures=indexes(X,y)

print(measures$internal)
print(measures$external)
}
