% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{multivariate_charac}
\alias{multivariate_charac}
\title{Multivariate characterization of the clustering}
\usage{
multivariate_charac(X, y, y_true = NULL, metric = "euclidean")
}
\arguments{
\item{X}{data.frame with the predictive variable that were used for the clustering}

\item{y}{data.frame (or vector) with the same number of rows (or length) as \code{X}. This represents the predicted classes.}

\item{y_true}{data.frame (or vector) with the same number of rows (or length) as \code{X}. This represents the predicted classes.}

\item{metric}{Metric used for the distance matrix. Can be chosen among : "euclidean", "maximum", "manhattan", "canberra", "binary" and "minkowski". Default value is "euclidean".}
}
\value{
multivariate - object containing descriptors of the multivariate characterization.
}
\description{
Multivariate characterization of the clustering
}
\details{
multivariate objects have the following attributes :
\itemize{
\item{"x"}{data.frame of the predictive variables.}
\item{"y"}{data.frame or vector of the clustering.}
\item{"y_true"}{(Optionnal) data.frame or vector of the true clustering.}
\item{"clusters_name"}{Vector of the names of the clusters.}
\item{"distance_matrix"}{Matrix of the distances between the center of clusters.}
\item{"correlation"}{List of the correlations of the whole data and in each cluster.}
\item{"squared_correlation"}{List of the squared correlations of the whole data and in each cluster.}
\item{"confusion_matrix"}{If y_true was provided, this is the confusionMatrix between the predicted clustering and the real one.}
}
}
\examples{
require(readxl)
df<- read_excel('../Autos.xlsx') #specify the path
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
## Creating a clustering using k-means with k=3
clus<- kmeans(X,3)
y <-clus$cluster

multivariate=multivariate_charac(X,y)
summary(multivariate)

}
