% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{Perform_UniChar}
\alias{Perform_UniChar}
\title{Class UniChar}
\usage{
Perform_UniChar(df, y)
}
\arguments{
\item{df}{a data frame representing all your data}

\item{y}{a data frame representing groups from a clustering process on numerical variables of df}
}
\value{
\item{instance}{object of UniChar class}
}
\description{
UniChar class creation.
To perform characterisation on a single variable after a clustering process.
}
\examples{
df<- read_excel('Autos.xlsx')
####Pre-processing####
#transform data
######################
y = kmeans(df[sep_data(df)[[2]]],4) #clustering example
Perform_UniChar(df,y)


}
