% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{plot.list.corr}
\alias{plot.list.corr}
\title{Plotting graphs to compare the correlation}
\usage{
\method{plot}{list.corr}(list_corr)
}
\arguments{
\item{list_corr}{list.corr object - List of the correlations of the whole data and in each cluster.}
}
\value{
None
}
\description{
Plotting the correlations in the whole dataset and within each cluster in order to find characteristic specific to clusters.
}
\examples{
require(readxl)
df<- read_excel('../Autos.xlsx') #specify the path
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
## Creating a clustering using k-means with k=3
clus<- kmeans(X,3)
y <-clus$cluster

multivariate=multivariate_charac(X,y)
plot(multivariate$correlation)

}
