% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{plot.measures}
\alias{plot.measures}
\title{Plotting measures of clustering}
\usage{
\method{plot}{measures}(measures)
}
\arguments{
\item{measures}{\itemize{
\item object of the class measures
}}
}
\value{
None
}
\description{
Plots a barplot with the internal indexes and, if y_true was given, a similar barplot for the external indexes
}
\examples{
require(readxl)
require(stats)
df<- read_excel('Autos.xlsx')
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
clus<- kmeans(X,3)
y <-clus$cluster

measures=indexes(X,y)

plot(measures)
}
