% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{plot.UniChar}
\alias{plot.UniChar}
\title{Bidimensional plot of numerical features of obj}
\usage{
\method{plot}{UniChar}(obj)
}
\arguments{
\item{obj}{of UniChar class}
}
\value{
none
}
\description{
Bidimensional plot aplied on numerical features of obj.
Selecteting the two most correlated features.
}
\examples{
#Import data
df<- read_excel('Autos.xlsx')
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]

#Clustering
clus<- kmeans(X,3)
y <-clus$cluster

#Attribut de la classe UniChar
obj <- Perform_UniChar(df,y)

#Plot clustering
plot(obj)
}
