% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{print.confusion.matrix}
\alias{print.confusion.matrix}
\title{Print for confusion.matrix object}
\usage{
\method{print}{confusion.matrix}(cm)
}
\arguments{
\item{cm}{confusion.matrix object}
}
\value{
None
Prints an Error message if the true clustering was not given in \code{multivariate_charac()}.
Otherwise, it prints the confusion matrix using the same method as the class confusionMatrix.
}
\description{
Print for confusion.matrix object
}
\examples{
require(readxl)
df<- read_excel('../Autos.xlsx') #specify the path
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
clus<- kmeans(X,3)
y <-clus$cluster

multivariate=multivariate_charac(X,y)
print(multivariate$confusion_matrix)

}
