% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{radar}
\alias{radar}
\title{Radar plot}
\usage{
radar(ind)
}
\arguments{
\item{ind}{an indicator to represent graphically (ex: effect size, Cramer's V, etc.)}
}
\value{
graph
}
\description{
Radar plot the indicator
}
\examples{
#Import data
df<- read_excel('Autos.xlsx')
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]

#Clustering
clus<- kmeans(X,3)
y <-clus$cluster

#Attribut de la classe UniChar
obj <- Perform_UniChar(df,y)

#Effect size
vt <- valuetest(obj, y)

#Plot radar
radar(vt)
}
