% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{summary.multivariate}
\alias{summary.multivariate}
\title{Summary of the object multivariate}
\usage{
\method{summary}{multivariate}(multivariate)
}
\arguments{
\item{multivariate}{object obtained as the output of \code{multivariate_charac}}
}
\value{
None
}
\description{
Prints various attributes and plot the correlation of multivariate.
}
\examples{
require(readxl)
df<- read_excel('../Autos.xlsx') #specify the path
X<-df[,c('length','width','height','engine-size','compression-ratio','horsepower','city-mpg','highway-mpg')]
## Creating a clustering using k-means with k=3
clus<- kmeans(X,3)
y <-clus$cluster

multivariate=multivariate_charac(X,y)
summarize(multivariate)

}
