% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniChar.R
\name{vcramer.UniChar}
\alias{vcramer.UniChar}
\title{Cramer's v}
\usage{
vcramer.UniChar(obj, y, label = NULL)
}
\arguments{
\item{obj}{of UniChar class}

\item{y}{a data frame representing groups from a clustering process on obj$data}

\item{label=NULL}{a specific label variable if you don't want all Cramer's v (can be several)}
}
\value{
Cramer's v value(s) according to your groups y on your categorical data
}
\description{
Calculate Cramer's v values on all yout categorical data or just on labeled ones if provided
}
\examples{
data <- read_excel('Autos.xlsx')
obj <- Perform_UniChar(data)
clus<- kmeans(obj$num,4)
group <-clus$cluster
y <- as.data.frame(group)
vcramer.UniChar(obj,y)

}
